requirements.txt
pygame-ce
PyQt-Fluent-Widgets[full]
svgwrite
numpy